requirements.txt
Flask==3.0.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
requests==2.32.3
Werkzeug==3.0.3
APScheduler==3.10.4
pywebpush==2.0.3
cryptography==42.0.8
python-dateutil==2.9.0.post0
